extract definition and paint to own modules
